+++ index.php
@@ -26,10 +26,10 @@
 require_once('../../config.php');
 require_once($CFG->dirroot . '/local/contact/class/local_contact.php');
 
-if (isset($_SERVER['HTTP_REFERER'])) {
-    $PAGE->set_url(get_local_referer(false));
-} else {
+if (empty(get_local_referer(false))) {
     $PAGE->set_url('/local/contact/index.php');
+} else {
+    $PAGE->set_url(get_local_referer(false));
 }
 
 // If we require user to be logged in.