+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Updated
 - Leading or trailing spaces in email addresses in recipient list settings will no longer cause a problem.
 - Fixed bug where forms could be submitted without sesskey JavaScript in the form.
+- Fixed missing reply-to address.
 
 ## [0.8.1] - 2017-11-13
 ### Updated