+++ CHANGELOG.md
@@ -1,13 +1,14 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## [0.8.2] - 2017-11-30
+## [0.8.2] - 2017-12-04
 ### Added
 - Settings option to require users to be logged-in when submitting the form. Not enforced by default. (Thanks to contributions by @kmoouni and @jezhops)
 ### Updated
 - Leading or trailing spaces in email addresses in recipient list settings will no longer cause a problem.
 - Fixed bug where forms could be submitted without sesskey JavaScript in the form.
 - Fixed missing reply-to address.
+- Fixed support for multiple selections for select tag.
 
 ## [0.8.1] - 2017-11-13
 ### Updated